# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.AOCRequest:
    properties:
      accountTaxId:
        type: integer
      accountTypeId:
        type: integer
      code:
        type: string
      description:
        maxLength: 255
        type: string
      id:
        type: string
      name:
        type: string
    required:
    - accountTaxId
    - accountTypeId
    - code
    - id
    - name
    type: object
  domain.AOCResponse:
    properties:
      accountTaxId:
        type: integer
      accountTypeId:
        type: integer
      code:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  domain.AccountTax:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      rate:
        type: number
      updatedAt:
        type: string
    type: object
...
  domain.H:
    additionalProperties: true
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@acareca.com
    name: API Support
  description: API documentation for ACA RECA Backend
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ACA RECA API
  version: "1.0"
paths:
  /coa:
    get:
      description: Get all chart of accounts entries
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.AOCResponse'
            type: array
      summary: Get all AOCs
      tags:
      - AOC
    post:
      consumes:
      - application/json
      description: Create a new coa with the given information
      parameters:
      - description: AOC information
        in: body
        name: coa
        required: true
        schema:
          $ref: '#/definitions/domain.AOCRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.AOCResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.H'
      summary: Create a new coa
      tags:
      - AOC
  /coa/account-types:
    get:
      description: 'Get all accounts. Query params: sort (code|name, default code),
        order (asc|desc, default asc)'
      parameters:
      - description: 'Sort field: code or name'
        in: query
        name: sort
        type: string
      - description: 'Sort order: asc or desc'
        in: query
        name: order
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.AOCResponse'
            type: array
      summary: Get all accounts (optionally sorted)
      tags:
      - AOC
  /coa/tax:
    get:
      consumes:
      - application/json
      description: Get all account tax types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AccountTax'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.H'
      summary: Get all account tax types
      tags:
      - AOC
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"